requirements.txt
# Core API & Web
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.34.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1

# Gemini LLM
google-generativeai==0.7.2

# Embeddings & Vector Search
sentence-transformers==3.0.1
faiss-cpu>=1.8.0

# BM25 Lexical Search
rank-bm25==0.2.2

# Re-ranking
# cross-encoder loaded via sentence-transformers

# Document Parsing
pypdf==4.2.0
python-docx==1.1.2

# Utilities
numpy==1.26.4
scikit-learn==1.5.0
tqdm==4.66.4
